requirements.txt
dbus-python==1.2.8
PyQt5==5.12.3